+++ tests/testthat/test_mlr_callbacks.R
@@ -34,6 +34,19 @@ test_that("backup callback works with standalone tuner", {
   expect_benchmark_result(readRDS(file))
 })
 
+test_that("backup callback errors when benchmark result is not stored", {
+  expect_error(tune(
+    tuner = tnr("random_search", batch_size = 2),
+    task = tsk("sonar"),
+    learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE)),
+    resampling = rsmp("cv", folds = 3),
+    measures = msr("classif.ce"),
+    term_evals = 4,
+    store_benchmark_result = FALSE,
+    callbacks = clbk("mlr3tuning.backup", path = tempfile(fileext = ".rds"))
+  ), "store_benchmark_result")
+})
+
 # batch measures callback ------------------------------------------------------
 
 test_that("batch measures callback works", {